Feature merge 1

+++ src/models/AccommodationModel.php
@@ -10,7 +10,7 @@ class AccommodationModel
         $this->connection = $connection;
     }
 
-    public function getByIdAndUserId(string $id, string $userId): array
+    public function getByIdAndUserId(int $id, int $userId): array
     {
         try {
             $stmt = $this->connection->prepare('SELECT * FROM accommodation WHERE id = ? AND user_id = ?');
@@ -23,12 +23,12 @@ class AccommodationModel
             return $data;
 
         } catch (PDOException $exception) {
-            error_log('AccommodationModel: getById: ' . $exception->getMessage() . 'id: ' . $id . 'userId: ' . $userId);
+            error_log('AccommodationModel: getByIdAndUserId: ' . $exception->getMessage() . 'id: ' . $id . 'userId: ' . $userId);
             throw $exception;
         }
     }
 
-    public function getAllByUserId(string $userId): array
+    public function getAllByUserId(int $userId): array
     {
         try {
             $stmt = $this->connection->prepare('SELECT * FROM accommodation WHERE user_id = ?');
@@ -46,21 +46,44 @@ class AccommodationModel
         }
     }
 
-    public function insert(int $userId, string $startDate, string $endDate, int $collegeId, string $purpose, string $supportingDocs): bool
+    public function getDocsNameByIdUserId(int $id, int $userId): array
     {
         try {
-            $sql = 'INSERT INTO accommodation (user_id, start_date, end_date, college_id, purpose, supporting_docs, status) 
-                    VALUES (:user_id, :start_date, :end_date, :college_id, :purpose, :supporting_docs, :status)';
+            $stmt = $this->connection->prepare('SELECT supporting_docs FROM accommodation WHERE id = ? AND user_id = ?');
+            $stmt->execute([$id, $userId]);
+            $data = $stmt->fetch();
+
+            if (!$data) {
+                return array();
+            }
+            return $data;
+
+        } catch (PDOException $exception) {
+            error_log('AccommodationModel: getDocsNameByIdUserId: ' . $exception->getMessage() . 'id: ' . $id . 'userId: ' . $userId);
+            throw $exception;
+        }
+    }
+
+    public function insert(int $userId, string $checkInDate, string $checkOutDate, int $collegeId, string $purpose, ?string $supportingDocs): string
+    {
+        try {
+            $sql = 'INSERT INTO accommodation (user_id, check_in_date, check_out_date, college_id, purpose, supporting_docs) 
+                    VALUES (:user_id, :check_in_date, :check_out_date, :college_id, :purpose, :supporting_docs)';
             $stmt = $this->connection->prepare($sql);
-            return $stmt->execute([
+            $success = $stmt->execute([
                 ':user_id' => $userId,
-                ':start_date' => $startDate,
-                ':end_date' => $endDate,
+                ':check_in_date' => $checkInDate,
+                ':check_out_date' => $checkOutDate,
                 ':college_id' => $collegeId,
                 ':purpose' => $purpose,
                 ':supporting_docs' => $supportingDocs,
-                ':status' => STATUS_PENDING
             ]);
+
+            if (!$success) {
+                return '';
+            }
+
+            return $this->connection->lastInsertId();
 
         } catch (PDOException $exception) {
             error_log('AccommodationModel: insert: ' . $exception->getMessage() . 'userId: ' . $userId);
@@ -68,21 +91,43 @@ class AccommodationModel
         }
     }
 
-    public function update(int $id, int $userId, string $startDate, string $endDate, int $collegeId, string $purpose, string $supportingDocs): bool
+    public function update(int $id, int $userId, string $checkInDate, string $checkOutDate, int $collegeId, string $purpose, ?string $supportingDocs, bool $deletePrevious): bool
     {
         try {
-            $sql = 'UPDATE accommodation SET start_date = :start_date, end_date = :end_date, college_id = :college_id, purpose = :purpose, supporting_docs = :supporting_docs 
-                    WHERE id = :id AND user_id = :user_id';
+
+            // delete uploaded file is user specifies or user uploaded new file
+            if ($deletePrevious || $supportingDocs != null){
+                $data = $this->getDocsNameByIdUserId($id, $userId);
+                if (empty($data)){
+                    // empty means record does not exist
+                    return false;
+                }
+
+                // delete uploaded file
+                if (!empty($data['supporting_docs']) && file_exists(ACCOMMODATION_UPLOAD_PATH.$data['supporting_docs'])){
+                    unlink(ACCOMMODATION_UPLOAD_PATH.$data['supporting_docs']);
+                }
+            }
+
+            $sql = 'UPDATE accommodation SET check_in_date = :check_in_date, check_out_date = :check_out_date, college_id = :college_id, purpose = :purpose, supporting_docs = :supporting_docs 
+                    WHERE id = :id AND user_id = :user_id AND status = :status';
             $stmt = $this->connection->prepare($sql);
-            return $stmt->execute([
+            $success = $stmt->execute([
                 ':id' => $id,
                 ':user_id' => $userId,
-                ':start_date' => $startDate,
-                ':end_date' => $endDate,
+                ':check_in_date' => $checkInDate,
+                ':check_out_date' => $checkOutDate,
                 ':college_id' => $collegeId,
                 ':purpose' => $purpose,
                 ':supporting_docs' => $supportingDocs,
+                ':status' => STATUS_SUBMITTED,
             ]);
+
+            if ($success && $stmt->rowCount() != 0) {
+                return true;
+            }
+
+            return false;
 
         } catch (PDOException $exception) {
             error_log('AccommodationModel: update: ' . $exception->getMessage() . 'id: ' . $id . 'userId: ' . $userId);
@@ -93,6 +138,17 @@ class AccommodationModel
     public function delete(int $id, int $userId): bool
     {
         try {
+            $data = $this->getDocsNameByIdUserId($id, $userId);
+            if (empty($data)){
+                // empty means record does not exist
+                return false;
+            }
+
+            // delete uploaded file
+            if (!empty($data['supporting_docs']) && file_exists(ACCOMMODATION_UPLOAD_PATH.$data['supporting_docs'])){
+                unlink(ACCOMMODATION_UPLOAD_PATH.$data['supporting_docs']);
+            }
+
             $stmt = $this->connection->prepare('DELETE FROM accommodation WHERE id = ? AND user_id = ?');
             return $stmt->execute([$id, $userId]);
